+++ semantic_frame/__init__.py
@@ -31,5 +31,5 @@ For Advanced Tool Use:
 
 from semantic_frame.main import describe_dataframe, describe_series
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 __all__ = ["describe_series", "describe_dataframe", "__version__"]